FileWriter improvements: part 3, file names
Now we can set not filename, but the pattern for it: prefix and suffix,
between which start time is inserted automatically. This allows to
easily re-create file when reopenning protocol (implemented this
behavior in MainWindow), and make these values persistent (they are now
stored in settigns).

In addition to re-creating the file, the plot is also re-created
(history is cleared) when closing protocol.

Also some re-structuring of mainwindow.cpp: introduced some slots (in
some cases lambda is not OK), used more range-for.